+++ src/__tests__/integration/Client-CreateProcessInstanceWithResult.spec.ts
@@ -1,5 +1,5 @@
-import { ZBClient } from '../../..'
-import { createUniqueTaskType } from '../../../lib/createUniqueTaskType'
+import { ZBClient } from '../..'
+import { createUniqueTaskType } from '../../lib/createUniqueTaskType'
 
 process.env.ZEEBE_NODE_LOG_LEVEL = process.env.ZEEBE_NODE_LOG_LEVEL || 'NONE'
 jest.setTimeout(25000)
