+++ src/__tests__/integration/Client-DeployProcess.spec.ts
@@ -1,5 +1,5 @@
-import { ZBClient } from '../../../index'
-import { createUniqueTaskType } from '../../../lib/createUniqueTaskType'
+import { ZBClient } from '../../index'
+import { createUniqueTaskType } from '../../lib/createUniqueTaskType'
 process.env.ZEEBE_NODE_LOG_LEVEL = process.env.ZEEBE_NODE_LOG_LEVEL || 'NONE'
 jest.setTimeout(20000)
 test('deploys a process', async () => {
