+++ src/__tests__/integration/Client-startProcess.spec.ts
@@ -1,6 +1,6 @@
-import { ZBClient } from '../../..'
-import { createUniqueTaskType } from '../../../lib/createUniqueTaskType'
-import { CreateProcessInstanceResponse } from '../../../lib/interfaces-grpc-1.0'
+import { ZBClient } from '../..'
+import { createUniqueTaskType } from '../../lib/createUniqueTaskType'
+import { CreateProcessInstanceResponse } from '../../lib/interfaces-grpc-1.0'
 
 process.env.ZEEBE_NODE_LOG_LEVEL = process.env.ZEEBE_NODE_LOG_LEVEL || 'NONE'
 jest.setTimeout(30000)
